+++ codex/librarian/watchdog/db_snapshot.py
@@ -14,12 +14,13 @@ class CodexDatabaseSnapshot(DirectorySnapshot, LoggerBaseMixin):
     """Take snapshots from the Codex database."""
 
     MODELS = (Folder, Comic, FailedImport, CustomCover)
+    COVERS_ONLY_MODELS = (CustomCover,)
     _STAT_LEN = 10
 
-    @classmethod
-    def _walk(cls, root):
+    def _walk(self, root):
         """Populate the DirectorySnapshot structures from the database."""
-        for model in cls.MODELS:
+        models = self.COVERS_ONLY_MODELS if self._covers_only else self.MODELS
+        for model in models:
             yield (
                 model.objects.filter(library__path=root)
                 .order_by("path")
@@ -57,7 +58,7 @@ class CodexDatabaseSnapshot(DirectorySnapshot, LoggerBaseMixin):
         i = (st.st_ino, st.st_dev)
         self._inode_to_path[i] = path
 
-    def __init__(
+    def __init__(  # noqa: PLR0913
         self,
         path,
         _recursive=True,  # unused, always recursive
@@ -65,8 +66,10 @@ class CodexDatabaseSnapshot(DirectorySnapshot, LoggerBaseMixin):
         _listdir=os.listdir,  # unused for database
         force=False,
         log_queue=None,
+        covers_only=False,
     ):
         """Initialize like DirectorySnapshot but use a database walk."""
+        self._covers_only = covers_only
         self.init_logger(log_queue)
         self._stat_info = {}
         self._inode_to_path = {}
@@ -81,10 +84,3 @@ class CodexDatabaseSnapshot(DirectorySnapshot, LoggerBaseMixin):
         for wp in chain.from_iterable(self._walk(path)):
             st = self._create_stat_from_db_stat(wp, stat, force)
             self._set_lookups(wp["path"], st)
-
-
-class CodexDatabaseCoverSnapshot(CodexDatabaseSnapshot):
-    """Take snapshots only for the custom cover dir."""
-
-    MODELS = (CustomCover,)
-    # Or an init variable on the other snapshot