+++ codex/librarian/watchdog/observers.py
@@ -21,8 +21,6 @@ class UatuMixin(BaseObserver, WorkerBaseMixin):
 
     ENABLE_FIELD = ""
     ALWAYS_WATCH = False
-    MODEL = Library
-    EVENT_HANDLER = CodexLibraryEventHandler
 
     def __init__(self, *args, **kwargs):
         """Initialize queues."""
@@ -55,7 +53,12 @@ class UatuMixin(BaseObserver, WorkerBaseMixin):
             return
 
         # Set up the watch
-        handler = self.EVENT_HANDLER(
+        handler_class = (
+            CodexCustomCoverEventHandler
+            if library.covers_only
+            else CodexLibraryEventHandler
+        )
+        handler = handler_class(
             library,
             librarian_queue=self.librarian_queue,
             log_queue=self.log_queue,
@@ -79,7 +82,7 @@ class UatuMixin(BaseObserver, WorkerBaseMixin):
     def sync_library_watches(self):
         """Watch or unwatch all libraries according to the db."""
         try:
-            libraries = self.MODEL.objects.all().only("pk", "path", self.ENABLE_FIELD)
+            libraries = Library.objects.all().only("pk", "path", self.ENABLE_FIELD)
             library_paths = set()
             for library in libraries:
                 try:
@@ -114,12 +117,14 @@ class UatuMixin(BaseObserver, WorkerBaseMixin):
 
             # If we don't have an emitter for this watch already, create it.
             if self._emitter_for_watch.get(watch) is None:
+                covers_only = isinstance(event_handler, CodexCustomCoverEventHandler)
                 emitter = self._emitter_class(
                     event_queue=self.event_queue,
                     watch=watch,
                     timeout=self.timeout,
                     log_queue=self.log_queue,
                     librarian_queue=self.librarian_queue,
+                    covers_only=covers_only
                 )
                 self._add_emitter(emitter)
                 if self.is_alive():
@@ -154,9 +159,7 @@ class LibraryPollingObserver(UatuMixin):
     def poll(self, library_pks, force=False):
         """Poll each requested emitter."""
         try:
-            qs = self.MODEL.objects
-            if library_pks:
-                qs = qs.filter(pk__in=library_pks)
+            qs = Library.objects.filter(pk__in=library_pks)
             paths = frozenset(qs.values_list("path", flat=True))
 
             for emitter in self.emitters:
@@ -175,15 +178,3 @@ class LibraryPollingObserver(UatuMixin):
         # shutdown event next.
         self.event_queue.put(self._SHUTDOWN_EVENT)
         super().on_thread_stop()
-
-
-class CustomCoverEventObserver(LibraryEventObserver):
-    """Custom Cover Dir Event Observer."""
-
-    EVENT_HANDLER = CodexCustomCoverEventHandler
-
-
-class CustomCoverPollingObserver(LibraryPollingObserver):
-    """Custom Cover Dir Polling Observer."""
-
-    EVENT_HANDLER = CodexCustomCoverEventHandler