+++ codex/librarian/watchdog/tasks.py
@@ -11,13 +11,6 @@ class WatchdogTask:
 
 
 @dataclass
-class WatchdogPollCustomCoversTask(WatchdogTask):
-    """Tell observer to poll the custom covers now."""
-
-    force: bool
-
-
-@dataclass
 class WatchdogPollLibrariesTask(WatchdogTask):
     """Tell observer to poll these libraries now."""