+++ codex/settings/settings.py
@@ -65,7 +65,7 @@ CODEX_PATH = BASE_DIR / "codex"
 CUSTOM_COVER_DIR = CONFIG_PATH / "custom-covers"
 CUSTOM_COVER_GROUP_DIRS = ("publishers", "imprints", "series", "volumes", "story-arcs")
 for group_dir in CUSTOM_COVER_GROUP_DIRS:
-    custom_cover_group_dir = CUSTOM_COVER_DIR/ group_dir
+    custom_cover_group_dir = CUSTOM_COVER_DIR / group_dir
     custom_cover_group_dir.mkdir(exist_ok=True, parents=True)
 
 # SECURITY WARNING: keep the secret key used in production secret!