+++ codex/views/admin/library.py
@@ -17,89 +17,50 @@ from codex.librarian.mp_queue import LIBRARIAN_QUEUE
 from codex.librarian.notifier.tasks import LIBRARY_CHANGED_TASK
 from codex.librarian.tasks import DelayedTasks
 from codex.librarian.watchdog.tasks import (
-    WatchdogPollCustomCoversTask,
     WatchdogPollLibrariesTask,
     WatchdogSyncTask,
 )
 from codex.logger.logging import get_logger
-from codex.models import CustomCoverDir, FailedImport, Folder, Library
+from codex.models import FailedImport, Folder, Library
 from codex.serializers.admin import (
     AdminFolderListSerializer,
     AdminFolderSerializer,
-    CustomCoverDirSerializer,
     FailedImportSerializer,
     LibrarySerializer,
 )
 
 LOG = get_logger(__name__)
 
+
 class AdminModelViewSet(ModelViewSet):
     """Admin ModelViewSet."""
 
     permission_classes: ClassVar[list] = [IsAdminUser]  # type: ignore
 
-class WatchdogChangeMixin:
-    """Common watchdog change methods."""
+
+class AdminLibraryViewSet(AdminModelViewSet):
+    """Admin Library Viewset."""
 
     WATCHDOG_SYNC_FIELDS = frozenset({"events", "poll", "pollEvery"})
 
+    queryset = Library.objects.prefetch_related("groups").defer(
+        "update_in_progress", "created_at", "updated_at"
+    )
+    serializer_class = LibrarySerializer
+
     @classmethod
     def _sync_watchdog(cls, validated_keys=None):
-        if validated_keys is None or validated_keys.intersection(cls.WATCHDOG_SYNC_FIELDS):
+        if validated_keys is None or validated_keys.intersection(
+            cls.WATCHDOG_SYNC_FIELDS
+        ):
             task = DelayedTasks(time() + 2, (WatchdogSyncTask(),))
             LIBRARIAN_QUEUE.put(task)
-
 
     @staticmethod
     def _on_change():
         cache.clear()
         task = LIBRARY_CHANGED_TASK
         LIBRARIAN_QUEUE.put(task)
-
-
-class AdminCustomCoverDirViewSet(AdminModelViewSet, WatchdogChangeMixin):
-    """Admin CustomCoverDir Viewset."""
-
-    queryset = CustomCoverDir.objects.defer(
-        "update_in_progress", "created_at", "updated_at"
-    )
-    serializer_class = CustomCoverDirSerializer
-    PK = 1
-
-    def get_object(self):
-        """Force pk to 1."""
-        if self.PK is not None:
-            self.kwargs["pk"] = self.PK
-        return super().get_object()
-
-    @staticmethod
-    def _poll(force):
-        task = WatchdogPollCustomCoversTask(force)
-        LIBRARIAN_QUEUE.put(task)
-
-    def perform_create(self, _serializer):
-        """Disallow."""
-        raise NotSupportedError
-
-    def perform_update(self, serializer):
-        """Perform update an run hooks."""
-        validated_keys = frozenset(serializer.validated_data.keys())
-        super().perform_update(serializer)
-        self._sync_watchdog(validated_keys)
-        self._poll(False)
-
-    def perform_destroy(self, _instance):
-        """Disallow."""
-        raise NotSupportedError
-
-
-class AdminLibraryViewSet(AdminModelViewSet, WatchdogChangeMixin):
-    """Admin Library Viewset."""
-
-    queryset = Library.objects.prefetch_related("groups").defer(
-        "update_in_progress", "created_at", "updated_at"
-    )
-    serializer_class = LibrarySerializer
 
     def _create_library_folder(self, library):
         folder = Folder(
@@ -118,6 +79,8 @@ class AdminLibraryViewSet(AdminModelViewSet, WatchdogChangeMixin):
     def perform_create(self, serializer):
         """Perform create and run hooks."""
         super().perform_create(serializer)
+        if serializer.validated_data.get("covers_only"):
+            raise NotSupportedError
         library = Library.objects.only("pk", "path").get(
             path=serializer.validated_data["path"]
         )
@@ -125,20 +88,23 @@ class AdminLibraryViewSet(AdminModelViewSet, WatchdogChangeMixin):
         self._sync_watchdog()
         self._poll(library.pk, False)
 
-
     def perform_update(self, serializer):
         """Perform update an run hooks."""
         validated_keys = frozenset(serializer.validated_data.keys())
+        pk = self.kwargs["pk"]
+        library = Library.objects.get(pk=pk)
+        if library.covers_only and serializer.validated_data.get("path"):
+            raise NotSupportedError
         super().perform_update(serializer)
         if "groupSet" in validated_keys:
             self._on_change()
         self._sync_watchdog(validated_keys)
-        pk = self.kwargs["pk"]
         self._poll(pk, False)
-
 
     def perform_destroy(self, instance):
         """Perform destroy and run hooks."""
+        if instance.covers_only:
+            raise NotSupportedError
         super().perform_destroy(instance)
         self._sync_watchdog()
         self._on_change()