+++ frontend/src/api/v3/admin.js
@@ -40,16 +40,6 @@ const deleteGroup = (pk) => {
   return HTTP.delete(`/admin/group/${pk}/`);
 };
 
-// Custom Cover Dir
-
-const getCustomCoversDirs = () => {
-  return HTTP.get("/admin/custom-cover-dir");
-};
-
-const updateCustomCoversDir = (pk, data) => {
-  return HTTP.put(`/admin/custom-cover-dir`, data);
-};
-
 // LIBRARIES
 
 const createLibrary = (data) => {
@@ -124,7 +114,6 @@ export default {
   deleteLibrary,
   deleteGroup,
   deleteUser,
-  getCustomCoversDirs,
   getFailedImports,
   getFlags,
   getFolders,
@@ -136,7 +125,6 @@ export default {
   librarianTask,
   postLibrarianTask,
   updateAPIKey,
-  updateCustomCoversDir,
   updateFlag,
   updateGroup,
   updateLibrary,