+++ frontend/src/components/admin/tabs/tabs.vue
@@ -22,14 +22,14 @@
         </v-window-item>
       </router-view>
     </v-window>
-    <div v-if="!librariesExist" id="noLibraries">
+    <div v-if="!comicLibrariesExist" id="noLibraries">
       Codex has no libraries. Select the Libraries tab and add a comic library.
     </div>
   </div>
 </template>
 
 <script>
-import { mapActions, mapGetters } from "pinia";
+import { mapActions, mapGetters, mapState } from "pinia";
 import titleize from "titleize";
 
 import { useAdminStore } from "@/stores/admin";
@@ -42,7 +42,10 @@ export default {
     };
   },
   computed: {
-    ...mapGetters(useAdminStore, ["librariesExist"]),
+    ...mapGetters(useAdminStore, ["comicLibrariesExist"]),
+    ...mapState(useAdminStore, {
+      librariesLoaded: (state) => Boolean(state.libraries),
+    }),
   },
   watch: {
     $route(to) {
@@ -52,7 +55,7 @@ export default {
     },
   },
   mounted() {
-    if (!this.librariesExist) {
+    if (!this.librariesLoaded) {
       this.loadTables(["Library"]);
     }
   },