+++ NEWS.md
@@ -12,8 +12,8 @@ border-radius: 128px;
   - Search one page of results at a time for speed.
   - Browser, Search, and Cover query optimizations.
   - Breadcrumbs in the browser
-  - Admin Flag to disable dynamic group covers in the browser in favor
-    of default group covers.
+  - Admin Flag to disable dynamic group covers in the browser in favor of
+    default group covers.
   - Smaller Browser controls.
   - Metadata tags can click to browse filtered on that tag.
   - Granular caching for covers in the browser.